Revert "Update dependencies, clean up code"
This reverts commit 90a92e6a34a720c3e29f621145c004c8513a306c.
The changes led to even worse conflicts with existing Java packages when
I tried to use this as an extension. I think I am going to need to
develop a version which is far more minimalist in scope, with far fewer
dependencies, for it to be safe as an extension.